+++ src/Repositories/SqlStatementRepository.php
@@ -41,77 +41,77 @@ class SqlStatementRepository extends \TechDivision\Import\Repositories\SqlStatem
      */
     private $statements = array(
         SqlStatementKeys::CATEGORY =>
-            'SELECT * FROM catalog_category_entity WHERE entity_id = :entity_id',
+            'SELECT * FROM ${table:catalog_category_entity} WHERE entity_id = :entity_id',
         SqlStatementKeys::CATEGORY_BY_PATH =>
-            'SELECT * FROM catalog_category_entity WHERE path = :path',
+            'SELECT * FROM ${table:catalog_category_entity} WHERE path = :path',
         SqlStatementKeys::CATEGORY_DATETIMES =>
             'SELECT *
-               FROM catalog_category_entity_datetime
+               FROM ${table:catalog_category_entity_datetime}
               WHERE entity_id = :pk
                 AND store_id = :store_id',
         SqlStatementKeys::CATEGORY_DECIMALS =>
             'SELECT *
-               FROM catalog_category_entity_decimal
+               FROM ${table:catalog_category_entity_decimal}
               WHERE entity_id = :pk
                 AND store_id = :store_id',
         SqlStatementKeys::CATEGORY_INTS =>
             'SELECT *
-               FROM catalog_category_entity_int
+               FROM ${table:catalog_category_entity_int}
               WHERE entity_id = :pk
                 AND store_id = :store_id',
         SqlStatementKeys::CATEGORY_TEXTS =>
             'SELECT *
-              FROM catalog_category_entity_text
+              FROM ${table:catalog_category_entity_text}
              WHERE entity_id = :pk
                AND store_id = :store_id',
         SqlStatementKeys::CATEGORY_VARCHARS =>
             'SELECT *
-               FROM catalog_category_entity_varchar
+               FROM ${table:catalog_category_entity_varchar}
               WHERE entity_id = :pk
                 AND store_id = :store_id',
         SqlStatementKeys::CATEGORY_DATETIMES_BY_PK_AND_STORE_ID =>
             'SELECT t0.*,
                     t1.attribute_code
-               FROM catalog_category_entity_datetime t0
-         INNER JOIN eav_attribute t1
+               FROM ${table:catalog_category_entity_datetime} t0
+         INNER JOIN ${table:eav_attribute} t1
                  ON t1.attribute_id = t0.attribute_id
               WHERE t0.entity_id = :pk
                 AND t0.store_id = :store_id',
         SqlStatementKeys::CATEGORY_DECIMALS_BY_PK_AND_STORE_ID =>
             'SELECT t0.*,
                     t1.attribute_code
-               FROM catalog_category_entity_decimal t0
-         INNER JOIN eav_attribute t1
+               FROM ${table:catalog_category_entity_decimal} t0
+         INNER JOIN ${table:eav_attribute} t1
                  ON t1.attribute_id = t0.attribute_id
               WHERE t0.entity_id = :pk
                 AND t0.store_id = :store_id',
         SqlStatementKeys::CATEGORY_INTS_BY_PK_AND_STORE_ID =>
             'SELECT t0.*,
                     t1.attribute_code
-               FROM catalog_category_entity_int t0
-         INNER JOIN eav_attribute t1
+               FROM ${table:catalog_category_entity_int} t0
+         INNER JOIN ${table:eav_attribute} t1
                  ON t1.attribute_id = t0.attribute_id
               WHERE t0.entity_id = :pk
                 AND t0.store_id = :store_id',
         SqlStatementKeys::CATEGORY_TEXTS_BY_PK_AND_STORE_ID =>
             'SELECT t0.*,
                     t1.attribute_code
-              FROM catalog_category_entity_text t0
-         INNER JOIN eav_attribute t1
+               FROM ${table:catalog_category_entity_text} t0
+         INNER JOIN ${table:eav_attribute} t1
                  ON t1.attribute_id = t0.attribute_id
               WHERE t0.entity_id = :pk
                 AND t0.store_id = :store_id',
         SqlStatementKeys::CATEGORY_VARCHARS_BY_PK_AND_STORE_ID =>
             'SELECT t0.*,
                     t1.attribute_code
-               FROM catalog_category_entity_varchar t0
-         INNER JOIN eav_attribute t1
+               FROM ${table:catalog_category_entity_varchar} t0
+         INNER JOIN ${table:eav_attribute} t1
                  ON t1.attribute_id = t0.attribute_id
               WHERE t0.entity_id = :pk
                 AND t0.store_id = :store_id',
         SqlStatementKeys::CREATE_CATEGORY =>
             'INSERT
-               INTO catalog_category_entity
+               INTO ${table:catalog_category_entity}
                     (attribute_set_id,
                      parent_id,
                      created_at,
@@ -129,7 +129,7 @@ class SqlStatementRepository extends \TechDivision\Import\Repositories\SqlStatem
                      :level,
                      :children_count)',
         SqlStatementKeys::UPDATE_CATEGORY =>
-            'UPDATE catalog_category_entity
+            'UPDATE ${table:catalog_category_entity}
                 SET attribute_set_id = :attribute_set_id,
                     parent_id = :parent_id,
                     created_at = :created_at,
@@ -141,7 +141,7 @@ class SqlStatementRepository extends \TechDivision\Import\Repositories\SqlStatem
               WHERE entity_id = :entity_id',
         SqlStatementKeys::CREATE_CATEGORY_DATETIME =>
             'INSERT
-               INTO catalog_category_entity_datetime
+               INTO ${table:catalog_category_entity_datetime}
                     (entity_id,
                      attribute_id,
                      store_id,
@@ -151,7 +151,7 @@ class SqlStatementRepository extends \TechDivision\Import\Repositories\SqlStatem
                      :store_id,
                      :value)',
         SqlStatementKeys::UPDATE_CATEGORY_DATETIME =>
-            'UPDATE catalog_category_entity_datetime
+            'UPDATE ${table:catalog_category_entity_datetime}
                 SET entity_id = :entity_id,
                     attribute_id = :attribute_id,
                     store_id = :store_id,
@@ -159,7 +159,7 @@ class SqlStatementRepository extends \TechDivision\Import\Repositories\SqlStatem
               WHERE value_id = :value_id',
         SqlStatementKeys::CREATE_CATEGORY_DECIMAL =>
             'INSERT
-               INTO catalog_category_entity_decimal
+               INTO ${table:catalog_category_entity_decimal}
                     (entity_id,
                      attribute_id,
                      store_id,
@@ -169,7 +169,7 @@ class SqlStatementRepository extends \TechDivision\Import\Repositories\SqlStatem
                     :store_id,
                     :value)',
         SqlStatementKeys::UPDATE_CATEGORY_DECIMAL =>
-            'UPDATE catalog_category_entity_decimal
+            'UPDATE ${table:catalog_category_entity_decimal}
                 SET entity_id = :entity_id,
                     attribute_id = :attribute_id,
                     store_id = :store_id,
@@ -177,7 +177,7 @@ class SqlStatementRepository extends \TechDivision\Import\Repositories\SqlStatem
               WHERE value_id = :value_id',
         SqlStatementKeys::CREATE_CATEGORY_INT =>
             'INSERT
-               INTO catalog_category_entity_int
+               INTO ${table:catalog_category_entity_int}
                     (entity_id,
                      attribute_id,
                      store_id,
@@ -187,7 +187,7 @@ class SqlStatementRepository extends \TechDivision\Import\Repositories\SqlStatem
                      :store_id,
                      :value)',
         SqlStatementKeys::UPDATE_CATEGORY_INT =>
-            'UPDATE catalog_category_entity_int
+            'UPDATE ${table:catalog_category_entity_int}
                 SET entity_id = :entity_id,
                     attribute_id = :attribute_id,
                     store_id = :store_id,
@@ -195,7 +195,7 @@ class SqlStatementRepository extends \TechDivision\Import\Repositories\SqlStatem
               WHERE value_id = :value_id',
         SqlStatementKeys::CREATE_CATEGORY_VARCHAR =>
             'INSERT
-               INTO catalog_category_entity_varchar
+               INTO ${table:catalog_category_entity_varchar}
                     (entity_id,
                      attribute_id,
                      store_id,
@@ -205,7 +205,7 @@ class SqlStatementRepository extends \TechDivision\Import\Repositories\SqlStatem
                      :store_id,
                      :value)',
         SqlStatementKeys::UPDATE_CATEGORY_VARCHAR =>
-            'UPDATE catalog_category_entity_varchar
+            'UPDATE ${table:catalog_category_entity_varchar}
                 SET entity_id = :entity_id,
                     attribute_id = :attribute_id,
                     store_id = :store_id,
@@ -213,7 +213,7 @@ class SqlStatementRepository extends \TechDivision\Import\Repositories\SqlStatem
               WHERE value_id = :value_id',
         SqlStatementKeys::CREATE_CATEGORY_TEXT =>
             'INSERT
-                INTO catalog_category_entity_text
+                INTO ${table:catalog_category_entity_text}
                      (entity_id,
                       attribute_id,
                       store_id,
@@ -223,7 +223,7 @@ class SqlStatementRepository extends \TechDivision\Import\Repositories\SqlStatem
                      :store_id,
                      :value)',
         SqlStatementKeys::UPDATE_CATEGORY_TEXT =>
-            'UPDATE catalog_category_entity_text
+            'UPDATE ${table:catalog_category_entity_text}
                 SET entity_id = :entity_id,
                     attribute_id = :attribute_id,
                     store_id = :store_id,
@@ -231,24 +231,24 @@ class SqlStatementRepository extends \TechDivision\Import\Repositories\SqlStatem
               WHERE value_id = :value_id',
         SqlStatementKeys::DELETE_CATEGORY =>
             'DELETE
-               FROM catalog_category_entity
+               FROM ${table:catalog_category_entity}
               WHERE path = :path',
         SqlStatementKeys::CATEGORY_COUNT_CHILDREN =>
-            'SELECT COUNT(*) FROM catalog_category_entity WHERE path LIKE :path'
+            'SELECT COUNT(*) FROM ${table:catalog_category_entity} WHERE path LIKE :path'
     );
 
     /**
-     * Initialize the the SQL statements.
+     * Initializes the SQL statement repository with the primary key and table prefix utility.
+     *
+     * @param \IteratorAggregate<\TechDivision\Import\Utils\SqlCompilerInterface> $compilers The array with the compiler instances
      */
-    public function __construct()
+    public function __construct(\IteratorAggregate $compilers)
     {
 
-        // call the parent constructor
-        parent::__construct();
+        // pass primary key + table prefix utility to parent instance
+        parent::__construct($compilers);
 
-        // merge the class statements
-        foreach ($this->statements as $key => $statement) {
-            $this->preparedStatements[$key] = $statement;
-        }
+        // compile the SQL statements
+        $this->compile($this->statements);
     }
 }