+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# Version 17.0.0
+
+## Bugfixes
+
+* None
+
+## Features
+
+* Add techdivision/import-cli-simple#216
+* Switch to latest techdivision/import 15.* version as dependency
+
 # Version 16.0.1
 
 ## Bugfixes